modify block access and test implenment

+++ block.go
@@ -32,10 +32,6 @@ func (b *BlockINode) GetBlockINodePath() string {
 }
 
 func (b *BlockINode) Save() error {
-	if _, err := os.Stat(b.GetBlockPath()); err != nil {
-		return xerrors.Errorf("error in os.Stat: %w", os.Create)
-	}
-
 	file, err := os.OpenFile(b.GetBlockINodePath(), os.O_WRONLY|os.O_CREATE|os.O_TRUNC, 0666)
 	if err != nil {
 		return xerrors.Errorf("error in os.OpenFile: %w", os.Create)
@@ -56,6 +52,7 @@ func (b *BlockINode) Save() error {
 
 func CreateBlock(block *BlockINode, id uint64) (BlockINode, error) {
 	if id != 0 {
+		// check is parent block exist
 		_, err := os.Stat(block.GetBlockPath())
 		if err != nil {
 			return BlockINode{}, xerrors.New(("parent block path not exist"))
@@ -71,45 +68,50 @@ func CreateBlock(block *BlockINode, id uint64) (BlockINode, error) {
 
 	// create block folder
 	if err := os.Mkdir(newBlock.GetBlockPath(), 0755); err != nil {
-		return BlockINode{}, xerrors.Errorf("error in os.Mkdir: %w", os.Create)
+		return BlockINode{}, xerrors.Errorf("error in os.Mkdir: %w", err)
 	}
 
 	// create block inode info
 	if err := newBlock.Save(); err != nil {
-		return BlockINode{}, xerrors.Errorf("error in block.Save: %w", os.Create)
+		return BlockINode{}, xerrors.Errorf("error in block.Save: %w", err)
 	}
 
 	return newBlock, nil
 }
 
-func GetBlock(path string, id uint64) (BlockINode, error) {
+func GetBlock(user *User, id uint64) (BlockINode, error) {
 	block := BlockINode{
-		UserPath: path,
+		UserPath: user.GetUserPath(),
 		NodeID:   id,
 	}
 
 	file, err := os.Open(block.GetBlockINodePath())
 	if err != nil {
-		return BlockINode{}, xerrors.Errorf("error in os.Open: %w", os.Create)
+		return BlockINode{}, xerrors.Errorf("error in os.Open: %w", err)
 	}
 	defer file.Close()
 
 	b, err := ioutil.ReadAll(file)
 	if err != nil {
-		return BlockINode{}, xerrors.Errorf("error in ioutil.ReadAll: %w", os.Create)
+		return BlockINode{}, xerrors.Errorf("error in ioutil.ReadAll: %w", err)
 	}
 
 	if err := json.Unmarshal(b, &block); err != nil {
-		return BlockINode{}, xerrors.Errorf("error in json.Unmarshal: %w", os.Create)
+		return BlockINode{}, xerrors.Errorf("error in json.Unmarshal: %w", err)
 	}
 
 	return block, nil
 }
 
-func DelteBlock(block *BlockINode) error {
+func DeleteBlock(user *User, id uint64) error {
+	block, ok := user.BlockMap[id]
+	if !ok {
+		return xerrors.New("block not exist")
+	}
+
 	// remove folder
 	if err := os.RemoveAll(block.GetBlockPath()); err != nil {
-		return xerrors.Errorf("error in os.RemoveAll: %w", os.Create)
+		return xerrors.Errorf("error in os.RemoveAll: %w", err)
 	}
 
 	return nil
